+++ src/main/java/blackjack/domain/Participant.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import java.util.HashMap;
 import java.util.List;
 
 public abstract class Participant {
@@ -42,5 +43,9 @@ public abstract class Participant {
 
     public int calculateTotalScore() {
         return drawnCards.sumScore();
+    }
+
+    public GameResult judgeResult(List<Participant> players, Participant dealer) {
+        return new GameResult(new HashMap<>(), new HashMap<>());
     }
 }
